Update ProductsPage.js
changed the @type property of the ProductGroup to ItemList and added an @id property with a unique identifier for each ProductGroup. Additionally, I used the itemListOrder property with the value http://schema.org/ItemListOrderDescending. Finally, I modified the itemListElement property to be an array of ListItem objects, each with a position property and an item property set to the individual product. --PhindAI